refactor: extract nb_jit boilerplate into _numba_compat module
Consolidate the identical 19-line Numba compatibility shim repeated
across five modules into a single src/pantr/_numba_compat.py file.
Each consumer now imports nb_jit with one line instead of carrying
the boilerplate locally.
